+++ lib/glimmer-dsl-web/samples/hello/hello_style.rb
@@ -83,9 +83,8 @@ class StyledButton
     button {
       inner_text <= [button_model, :text, computed_by: :pushed]
       
-      class_name <= [button_model, :pushed,
-                      on_read: ->(pushed) { pushed ? 'pushed' : 'pulled' }
-                    ]
+      class_name(:pushed) <= [button_model, :pushed]
+      class_name(:pulled) <= [button_model, :pushed, on_read: :!]
       
       style(:width) <= [button_model, :width, on_read: :px]
       style(:height) <= [button_model, :height, on_read: :px]
@@ -99,22 +98,22 @@ class StyledButton
     }
   }
   
-  style {"
-    .#{component_element_class} {
-      font-family: Courrier New, Courrier;
-      border-radius: 5px;
-      border-width: 17px;
-      margin: 5px;
+  style {
+    r(component_element_selector) {
+      font_family 'Courrier New, Courrier'
+      border_radius 5
+      border_width 17
+      margin 5
     }
     
-    .#{component_element_class}.pulled {
-      border-style: outset;
+    r("#{component_element_selector}.pulled") {
+      border_style :outset
     }
     
-    .#{component_element_class}.pushed {
-      border-style: inset;
+    r("#{component_element_selector}.pushed") {
+      border_style :inset
     }
-  "}
+  }
 end
 
 class StyledButtonRangeInput
@@ -172,22 +171,22 @@ class HelloStyle
     }
   }
   
-  style {"
-    .styled-button-form {
-      padding: 20px;
-      display: inline-grid;
-      grid-template-columns: auto auto;
+  style {
+    r('.styled-button-form') {
+      padding 20
+      display 'inline-grid'
+      grid_template_columns 'auto auto'
     }
     
-    .styled-button-form label, input {
-      display: block;
-      margin: 5px 5px 5px 0;
+    r('.styled-button-form label, input') {
+      display :block
+      margin '5px 5px 5px 0'
     }
     
-    .#{component_element_class} .styled-button {
-      display: block;
+    r("#{component_element_selector} .styled-button") {
+      display :block
     }
-  "}
+  }
 end
 
 Document.ready? do