+++ lib/glimmer/data_binding/element_binding.rb
@@ -10,7 +10,9 @@ module Glimmer
       attr_reader :element, :property, :translator, :sub_property
       def initialize(element, property, translator: nil)
         @element = element
-        if (property_parts = property.to_s.match(Glimmer::Web::ElementProxy::REGEX_STYLE_SUB_PROPERTY))
+        if (property_parts = property.to_s.match(Glimmer::Web::ElementProxy::REGEX_CLASS_NAME_SUB_PROPERTY))
+          @property, @sub_property = property_parts.to_a.drop(1)
+        elsif (property_parts = property.to_s.match(Glimmer::Web::ElementProxy::REGEX_STYLE_SUB_PROPERTY))
           @property, @sub_property = property_parts.to_a.drop(1)
           @sub_property = @sub_property.gsub('_', '-')
         else
@@ -27,9 +29,11 @@ module Glimmer
       def call(value)
         evaluated_property_value = evaluate_property
         converted_value = @translator&.call(value, evaluated_property_value) || value
-        unless converted_value == evaluated_property_value
+        if converted_value != evaluated_property_value
           if @sub_property
-            if @property.to_s == 'style'
+            if @property.to_s == 'class_name'
+              @element.class_name_included(@sub_property, converted_value)
+            elsif @property.to_s == 'style'
               @element.style_property(@sub_property, converted_value)
             end
           else
@@ -40,7 +44,9 @@ module Glimmer
       
       def evaluate_property
         if @sub_property
-          if @property.to_s == 'style'
+          if @property.to_s == 'class_name'
+            @element.class_name_included(@sub_property)
+          elsif @property.to_s == 'style'
             @element.style_property(@sub_property)
           end
         else