+++ lib/glimmer/dsl/web/dsl.rb
@@ -28,6 +28,7 @@ require 'glimmer/dsl/web/a_expression'
 require 'glimmer/dsl/web/span_expression'
 require 'glimmer/dsl/web/style_element_expression'
 require 'glimmer/dsl/web/inline_style_data_binding_expression'
+require 'glimmer/dsl/web/class_name_inclusion_data_binding'
 require 'glimmer/dsl/web/bind_expression'
 require 'glimmer/dsl/web/data_binding_expression'
 require 'glimmer/dsl/web/content_data_binding_expression'
@@ -45,6 +46,7 @@ module Glimmer
          style_element
          content_data_binding
          inline_style_data_binding
+         class_name_inclusion_data_binding
          component
          formatting_element
          data_binding